In an attempt to make Google Maps work consistently, made it worse
App no longer displays line on the map. The solution is probably to save the promised city marker data from the database in the service instead of trying to get it per request and wait on it. My effort to deal with the promise caused me to make too many changes to fix them in the time that remains.